requirements.txt
torch==2.0.0
torchvision==0.15.1
numpy==1.24.2
tqdm==4.65.0
pillow==9.5.0
opencv-python==4.6.0.66 # some other versions do not work
# can import: 
# 4.6.0.66, 4.8.0.74

torchmetrics==0.11.4
scipy==1.10.1 # required by torchmetrics
torch-fidelity==0.3.0 # required by torchmetrics fid
facenet-pytorch==2.5.3
deepface==0.0.79 # uses tensorflow, keras, flask, and lots of other stuff
sixdrepnet==0.1.6 # face pose evaluation
evotorch==0.4.1

tensorboard==2.12.0
matplotlib==3.7.1
pandas==1.5.3
tabulate==0.9.0 # required by pandas
seaborn==0.12.2

scikit-image==0.21.0 # to load images as a dataset